+++ .idea/gradle.xml
@@ -10,6 +10,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/annotation" />
+            <option value="$PROJECT_DIR$/consumer" />
             <option value="$PROJECT_DIR$/processor" />
           </set>
         </option>
