+++ annotation/build.gradle.kts
@@ -5,9 +5,17 @@ plugins {
 group = "org.example"
 version = "1.0-SNAPSHOT"
 
-repositories {
-    mavenCentral()
+
+java {
+    toolchain.languageVersion.set(JavaLanguageVersion.of(8))
 }
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+
 
 dependencies {
     implementation(kotlin("stdlib"))
