+++ consumer/build.gradle.kts
@@ -1,14 +1,23 @@
 plugins {
-    kotlin("jvm") version "1.5.10"
+    kotlin("jvm")
+    id("com.google.devtools.ksp") version "1.5.31-1.0.0"
 }
 
 group = "org.example"
 version = "1.0-SNAPSHOT"
 
-repositories {
-    mavenCentral()
-}
-
 dependencies {
     implementation(kotlin("stdlib"))
+    implementation(project(":annotation"))
+    ksp(project(":processor"))
+}
+
+java {
+    toolchain.languageVersion.set(JavaLanguageVersion.of(8))
+}
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
 }
