+++ consumer/src/main/kotlin/com/tsongkha/kspexample/consumer/Main.kt
@@ -1,0 +1,19 @@
+package com.tsongkha.kspexample.consumer
+
+import com.tsongkha.kspexample.annotation.IntSummable
+
+fun main() {
+    println("hello")
+    val foo = Foo(1,2)
+//    foo.sumInts
+}
+
+data class Hello(
+    val int: Int
+)
+
+@IntSummable
+data class Foo(
+    val x: Int,
+    val y: Int
+)
