+++ processor/src/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,0 +1,1 @@
+com.tsongkha.kspexample.processor.IntSummableProcessor.IntSummableProcessorProvider
