--- src/main/kotlin/com/tsongkha/kspexample/Foo.kt
@@ -1,9 +1,0 @@
-package com.tsongkha.kspexample
-
-import com.tsongkha.kspexample.annotation.IntSummable
-
-@IntSummable
-data class Foo(
-    val x: Int,
-    val y: Int
-)
