--- src/main/kotlin/com/tsongkha/kspexample/Main.kt
@@ -1,6 +1,0 @@
-package com.tsongkha.kspexample
-
-fun main() {
-    println("hello")
-    val foo = Foo(1,2).sumInts()
-}
